- added the new Switcher component - added switcher in mobile menu - added two variables - refactor - added to do
