+++ README.md
@@ -1,1 +1,6 @@
 # Responsive layout in development
+
+## to do:
+
+- Fix grid display in adaptive layout
+- Fix routing in adaptive layout for display of the components
