+++ src/components/Header/CustomNickName/CustomNickName.module.scss
@@ -14,7 +14,7 @@
     border: initial;
     color: $main_color;
 
-    @media screen and (max-width: 928px) {
+    @media screen and (max-width: $tablet_width) {
       @include flex(flex-start, center);
     }
 
@@ -47,7 +47,7 @@
       padding: 5px 20px;
     }
 
-    @media screen and (max-width: 928px) {
+    @media screen and (max-width: $tablet_width) {
       button {
         font-size: 12px;
         padding: 1px 12px;
@@ -66,7 +66,7 @@
     padding: 5px;
     margin-bottom: 10px;
 
-    @media screen and (max-width: 928px) {
+    @media screen and (max-width: $tablet_width) {
       padding: 4px;
     }
 
@@ -84,7 +84,7 @@
   label {
     margin-bottom: 10px;
 
-    @media screen and (max-width: 928px) {
+    @media screen and (max-width: $tablet_width) {
       margin-bottom: 5px;
     }
   }
