+++ src/components/Header/HeaderLogo.tsx
@@ -1,13 +1,15 @@
-import { FC }          from 'react'
-import s               from './Header.module.scss'
-import { Link }        from 'react-router-dom'
-import { THeaderLogo } from './types'
+import { FC }   from 'react'
+import s        from './Header.module.scss'
+import { Link } from 'react-router-dom'
 
-import logoLight from '../../assets/images/main/logoLight.webp'
-import logoDark  from '../../assets/images/main/logoDark.webp'
+import logoLight       from '../../assets/images/main/logoLight.webp'
+import logoDark        from '../../assets/images/main/logoDark.webp'
+import { useSelector } from 'react-redux'
+import { themeFilter } from '../../redux/theme/selectors'
 
 
-const HeaderLogo: FC<THeaderLogo> = ({ theme }) => {
+const HeaderLogo: FC = () => {
+	const { theme } = useSelector(themeFilter)
 
 	return (
 		<Link className={ s.logo } to='/'>
