+++ src/components/MobileMenu/MobileMenu.module.scss
@@ -28,12 +28,14 @@
     @include flex(center);
     gap: 10px;
 
-    @media screen and (max-width: 560px) {
+    @media screen and (max-width: $mobileWidth) {
       @include flex(center, center, column);
     }
   }
 
   .userName {
+    @include flex(center, center);
+    gap: 20px;
     position: absolute;
     top: 30px;
     left: 20px;
