+++ src/components/MobileMenu/MobileMenu.tsx
@@ -9,6 +9,7 @@ import { useCardsData } from '../../hooks/useCardsData'
 import Card           from '../MainNavCards/Card'
 import { Close }      from '../Ui'
 import CustomNickName from '../Header/CustomNickName/CustomNickName'
+import ThemeSwitcher  from '../shared/ThemeSwitcher/ThemeSwitcher'
 
 
 const MobileMenu = () => {
@@ -32,7 +33,10 @@ const MobileMenu = () => {
 	return (
 		<div className={ `${ s.wrapper } ${ fadeIn ? s.animShow : '' }` }>
 			<Close onClick={ onCloseModal }/>
-			<div className={ s.userName }><CustomNickName/></div>
+			<div className={ s.userName }>
+				<CustomNickName/>
+				<ThemeSwitcher/>
+			</div>
 			<div className={ s.menuCards }>
 				{ cards.map((card) => <Card key={ card.name } onClick={ onCloseModal } { ...card }/>) }
 			</div>
