+++ src/components/Ui/UploadBtn/UploadBtn.module.scss
@@ -1,8 +1,10 @@
+@import "src/scss/variables";
+
 .wrapper {
   margin-right: 0;
   margin-left: auto;
 
-  @media screen and (max-width: 560px) {
+  @media screen and (max-width: $mobileWidth) {
     width: 100%;
     margin-right: 0;
     margin-left: 0;
