+++ src/components/UploadModal/UploadModal.module.scss
@@ -21,7 +21,7 @@
     @include transition();
     transform: translateX(0);
   }
-  
+
   .header {
     margin-top: 80px;
     @include flex(null, center, column);
@@ -32,12 +32,12 @@
       font-size: 36px;
       line-height: 52px;
 
-      @media screen and (max-width: 928px) {
+      @media screen and (max-width: $tablet_width) {
         font-size: 32px;
         line-height: 48px;
       }
 
-      @media screen and (max-width: 560px) {
+      @media screen and (max-width: $mobileWidth) {
         font-size: 20px;
         line-height: 29px;
         margin-bottom: 20px;
@@ -49,12 +49,12 @@
       font-size: 20px;
       line-height: 30px;
 
-      @media screen and (max-width: 928px) {
+      @media screen and (max-width: $tablet_width) {
         font-size: 18px;
         line-height: 26px;
       }
 
-      @media screen and (max-width: 928px) {
+      @media screen and (max-width: $tablet_width) {
         font-size: 20px;
         line-height: 30px;
       }
