+++ src/components/Voting/NavButtons/NavButtons.module.scss
@@ -11,7 +11,7 @@
   border: 4px solid var(--bg_white-grey);
   background: white;
 
-  @media screen and (max-width: 560px) {
+  @media screen and (max-width: $mobileWidth) {
     width: 186px;
     height: 60px;
   }
@@ -45,14 +45,14 @@
     svg {
       fill: white;
 
-      @media screen and (max-width: 560px) {
+      @media screen and (max-width: $mobileWidth) {
         width: 22.5px;
         height: 22.5px;
       }
     }
 
     .heartImg {
-      @media screen and (max-width: 560px) {
+      @media screen and (max-width: $mobileWidth) {
         width: 22.5px;
         height: 19.5px;
       }
