+++ src/components/Voting/Voting.module.scss
@@ -8,7 +8,7 @@
   margin-bottom: 52px;
   position: relative;
 
-  @media screen and (max-width: 560px) {
+  @media screen and (max-width: $mobileWidth) {
     height: 206px;
     margin: 0 auto;
   }
