+++ src/components/layouts/ContentBody/ContentBody.module.scss
@@ -1,3 +1,5 @@
+@import "src/scss/variables";
+
 .wrapper {
   width: 100%;
 
@@ -14,7 +16,7 @@
     opacity: 0;
     transition: transform 1s ease;
 
-    @media screen and (max-width: 567px) {
+    @media screen and (max-width: $mobileWidth) {
       height: auto;
     }
