+++ src/components/layouts/Desktop/DesktopLayout.module.scss
@@ -1,4 +1,4 @@
-@import "../../../scss/mixins";
+@import "src/scss/variables";
 
 .wrapper {
   @include flex(space-between);
@@ -12,7 +12,7 @@
     padding: 30px 30px 0 30px;
   }
 
-  @media screen and (max-width: 928px) {
+  @media screen and (max-width: $tablet_width) {
     padding: 30px 15px 0 15px;
   }
 }
