+++ src/components/layouts/MainNavigation/MainNavigation.tsx
@@ -1,26 +1,14 @@
-import { FC }         from 'react'
-import s              from './MainNavigation.module.scss'
-import { setLSTheme } from '../../../services/localStorage/theme'
-
-import { useSelector }    from 'react-redux'
-import { useAppDispatch } from '../../../redux/store'
-import { themeFilter }    from '../../../redux/theme/selectors'
+import { FC } from 'react'
+import s      from './MainNavigation.module.scss'
 
 import { Header, MainNavCards } from '../../index'
 
 
 const MainNavigation: FC = () => {
-	const { theme } = useSelector(themeFilter)
-	const dispatch = useAppDispatch()
-
-	const onChangeTheme = () => {
-		setLSTheme(theme, dispatch)
-	}
-
 	return (
 		<div className={ s.wrapper }>
 			<div className={ s.container }>
-				<Header theme={ theme } onChangeTheme={ onChangeTheme }/>
+				<Header/>
 				<MainNavCards/>
 			</div>
 		</div>
