+++ src/components/shared/Pagination/Pagination.module.scss
@@ -14,7 +14,7 @@
     text-align: left;
   }
 
-  @media screen and (max-width: 560px) {
+  @media screen and (max-width: $mobileWidth) {
     .prev, .next {
       button {
         @include flex(center, center);
