+++ src/components/shared/ThemeSwitcher/ThemeSwitcher.module.scss
@@ -1,0 +1,73 @@
+@import "../../../scss/variables";
+
+.navigation {
+  grid-area: switcher;
+  @include flex(flex-end, center);
+
+  &:focus-visible {
+    outline: 2px solid $main_color !important;
+  }
+
+  .imgWrapper {
+    background: var(--bg_white-grey);
+    width: 24px;
+    height: 24px;
+    border-radius: 50px;
+    @include flex(center, center);
+    @include transition();
+  }
+
+  .switcher {
+    background: rgba(255, 134, 142, 0.2);
+    border: none;
+    border-radius: 50px;
+    width: 44px;
+    height: 24px;
+    margin-left: 10px;
+    position: relative;
+    cursor: pointer;
+    z-index: 5;
+
+    button {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      background: transparent;
+      z-index: 6;
+      border: none;
+      border-radius: 10px;
+      cursor: pointer;
+    }
+
+    input {
+      -webkit-appearance: none;
+      appearance: none;
+      position: absolute;
+      z-index: -1;
+    }
+
+    span::after {
+      content: '';
+      position: absolute;
+      display: block;
+      width: 16px;
+      height: 16px;
+      background: #FF868E;
+      border-radius: 50%;
+      right: 0;
+      top: 4px;
+      transform: translateX(-5px);
+      @include transition();
+      cursor: pointer;
+      z-index: 2;
+    }
+
+    input:checked ~ span::after {
+      transform: translateX(-24px);
+      @include transition();
+    }
+  }
+
+}
